% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_image.R
\name{upload_image}
\alias{upload_image}
\title{Upload issue images to GitHub}
\usage{
upload_image(
  file = NULL,
  caption = NULL,
  width = "600",
  issue = NULL,
  image_repo = NULL,
  push = TRUE
)
}
\arguments{
\item{file}{path to image which is to be uploaded}

\item{caption}{caption to insert into markdown link}

\item{width}{maximum width of resized image}

\item{issue}{current issue being edited (will be inferred from draft.md if needed)}

\item{image_repo}{local path to rweekly/image repository}

\item{push}{should the data be pushed to the repo immediately? default: TRUE}
}
\value{
the markdown link (invisibly) after being printed to the console.
}
\description{
Resizes an image to a maximum width, uploads both files to GitHub,
then generates a link to the resized file in markdown format.
}
\details{
\code{file} will be resized to a maximum width of \code{width} pixels.
}
